pyproject.toml
[tool.poetry]
name = "freddy"
version = "1.0.0"
description = ""
authors = ["Harry <devharry20@github.com>"]
license = "MIT"

[tool.poetry.scripts]
freddy = "freddy.main:main"

[tool.poetry.dependencies]
python = "^3.9"
discord = "^1.0.1"
python-dotenv = "^0.17.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
